add multiple actions in card
