+++ lib/constants/domain/app_icons.dart
@@ -1,0 +1,7 @@
+class AppIcons {
+  static const String list = 'res/icons/list.svg';
+  static const String map = 'res/icons/map.svg';
+  static const String favorite = 'res/icons/favorite.svg';
+  static const String settings = 'res/icons/settings.svg';
+  static const String heart = 'res/icons/heart.svg';
+}
