+++ lib/ui/screens/sight_list_screen.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
+import 'package:places/constants/domain/app_icons.dart';
 import 'package:places/domain/sight.dart';
 import 'package:places/ui/widgets/custom_app_bar.dart';
 import 'package:places/ui/widgets/sight_card.dart';
@@ -23,7 +25,12 @@ class _SightListScreenState extends State<SightListScreen> {
         padding: const EdgeInsets.all(10),
         child: Column(
           children: widget.sights.map((sight) {
-            return SightCard(sight: sight);
+            return SightCard(
+              sight: sight,
+              actions: [
+                SvgPicture.asset(AppIcons.heart),
+              ],
+            );
           }).toList(),
         ),
       ),
